requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
alembic==1.13.1
httpx==0.27.0

# PDF processing
pypdf==4.3.1
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.7
langchain-openai==0.2.14
langchain-chroma==0.1.4
langchain-text-splitters==0.3.2

# Vector DB + embeddings
chromadb==0.5.20
openai==1.55.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
